// app/interface/card-course.interface.tsx
export interface CardCourseInterface {
  image: string;
  title: string;
  percent: number;
  bgImage: string;
}

export interface CardCourseOnlineInterface {
  bgImage: string;
  tag: string;
  title: string;
  detail: string;
  profile: string;
  name: string;
}

export interface CardNewsInterface {
  title: string;
  image: string;
  bgImage: string;
  tag: string;
  subTitle: string;
  detail: string;
  date: string;
  view: string;
}

export interface CardNewsLatestInterface {
  title: string;
  bgImage: string;
  tag: string;
  subTitle: string;
  detail: string;
}